Define resources and methods.
With our user table in the database, we can now look at defining some resources
and methods to work with them. I've chosen 2 resources, the 'Accounts' resource
which represents a *list* of accounts, and the user resource which is for a
*single* account.

With these accounts, we can define some methods:

    * GET /user/:name - view an account (maybe a list of bookmarks, recent
      activity, etc)
    * PUT /user/:name - modify account details
    * DELETE /user/:name - delete an account

    * POST /accounts - register a new account

The separatation here might look a little strange at first, but remember POST
means to append to a resource. In this case, the user account is sub-ordinate to
our accounts resource, so appending a new account makes sense here.