Add a context to our application.
Our methods are going to touch the DB, so we need a convenient way to pass a
database handle around. As our schema is pretty straightforward, I'm going to
use DBIx::Class to abstract the database out.

I packaged this context up in the Slothmarks module, and instantiate it in our
REST server. This allows me to access the Slothmarks context outside the REST
server, for example I may need it later in cron jobs, or scripts.