Implement GET for accounts.
We are now ready to begin implementing our REST service!

The easiest thing to do would be start by getting user accounts from the DB. We
use DBIx::Simple to run our query, and then throw a 404 'NotFound' exception if
the user couldn't be found. In this case, it will always do that, because we
haven't implemented a method to register! Let's do that now...