Implement account creation.
As we mentioned before, a POST to /accounts should 'append' a new account to the
account list, which means registering a new account. We first check for an
already existing account with a given username -- if that's taken we return a
409 conflict error.

Otherwise, we insert the row, and then 'throw' the created message. This skips
any further serialization, and adds the Location header. We use
$request->uri_for to generate a URI based on various path components.