Implement GET for users
Now that we are able to register, we actually have some user objects that can be
fetched, but we haven't yet specified a way to serialize them.

I've added a basic JSON serializer, which is dead simple. We dispatch on the
object type we get, and in this case we serialize just the user name and full
name, respecting some privacy by not serializing everything.

Also, now that we can serialize accounts, I'm serializing the newly created
accounts in our account list POST handler. While the resource can be found by
following the Location header, this may save clients an extra request.