Allow submitting bookmarks
We now implement POST for the user's bookmarks resource. Again, we choose to
POST to a resource to append to it, in this case appending a new bookmark. We
first make sure the user hasn't already bookmarked the given URI by creating a
canonical form of it and seeing if it's been bookmarked.

If not, we create a new bookmark (find-or-create), and then create a new user
bookmark entity too. Finally, we return a response containing this bookmark, and
also the location where it can be found.

We extend the serializer to allow serializing user bookmark objects, in a fairly
trivial representation.